Fixed a bug when locating JavaScript runtime, and refactored common code
